requirements.txt
beautifulsoup4
pandas
numpy
requests
lxml
dash_bootstrap_components
dash
python-dotenv